♻️ Refactor: 그룹 선택 페이지 로직 분리
